extracted tileset config to json
